Renamed 'flush...' methods
